# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Webinar ApplicationAPI",
        "version": "0.1"
    },
    "paths": {
...
        "/api/search": {
            "post": {
                "tags": [
                    "search"
                ],
                "summary": "return search result",
                "description": "return search result.\n      isFree and hasDiscount must be 0 or 1.\n      isEvent could be null, 0 or 1",
                "operationId": "returnsearchresult",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "isFree",
                                    "hasDiscount"
                                ],
                                "properties": {
                                    "subject_id": {
                                        "type": "integer"
                                    },
                                    "isFree": {
                                        "type": "integer"
                                    },
                                    "hasDiscount": {
                                        "type": "integer"
                                    },
                                    "keyword": {
                                        "type": "string"
                                    },
                                    "isEvent": {
                                        "type": "integer"
                                    },
                                    "type_id": {
                                        "type": "integer"
                                    },
                                    "organizer_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/api/user/saveTransactions": {
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "save transactions in DB",
                "description": "save transactions in DB",
                "operationId": "transactions",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "transaction_id",
                                    "ref_id",
                                    "tracking_code",
                                    "ip",
                                    "code",
                                    "amount",
                                    "invoice_id",
                                    "gateway",
                                    "result_message",
                                    "status"
                                ],
                                "properties": {
                                    "transaction_id": {
                                        "type": "integer"
                                    },
                                    "ref_id": {
                                        "type": "text"
                                    },
                                    "tracking_code": {
                                        "type": "text"
                                    },
                                    "ip": {
                                        "type": "text"
                                    },
                                    "code": {
                                        "type": "integer"
                                    },
                                    "amount": {
                                        "type": "integer"
                                    },
                                    "invoice_id": {
                                        "type": "integer"
                                    },
                                    "gateway": {
                                        "type": "integer"
                                    },
                                    "result_message": {
                                        "type": "text"
                                    },
                                    "status": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP sent Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "security": {
        "passport": {
            "type": "oauth2",
            "description": "Laravel webinar passport oauth2 security.",
            "in": "header",
            "scheme": "https",
            "flows": {
                "password": {
                    "authorizationUrl": "http://localhost/oauth/authorize",
                    "tokenUrl": "http://localhost/oauth/token",
                    "refreshUrl": "http://localhost/token/refresh",
                    "scopes": []
                }
            }
        }
    }
}